Wait for svc context before sending RPC response
